==> Data analysis/01 - Load climate - ERA5.Rmd <==
---
title: "Load climate"
author: "Abby Lewis"
date: "2022-10-10"
output: html_document
---

This script loads climate data from the ERA5 re-analysis data product and matches climate data to LakeIDs by lat/lon. The raw ERA 5 data product is too large to store on github, so will need to be re-downloaded. If you would rather skip this step, the output of this code file (temperature and precipitation records at each lake) has been created and stored on github in "./External data/historical_temp_output_era5.csv" and "./External data/historical_precip_output_era5.csv"

To download ERA 5 temperature and precipitation data: https://cds.climate.copernicus.eu/cdsapp#!/dataset/reanalysis-era5-single-levels-monthly-means?tab=form
  Product type: Monthly averaged reanalysis
  Variable: "Total precipitation" and "2m temperature"
  Year: 1940-2022
  Month: all
  Time: 00:00
  Geographical Area: "Whole available region"
  Format: "NetCDF (experimental)"

Table of Contents:
- Step 1: Load packages
- Step 2: Load and synthesize data


Step 1: Load packages
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
Sys.setenv(TZ='UTC')
library(ncdf4)
library(ggplot2)
library(fields)
library(tidyverse)
library(data.table)
library(lubridate)
```


Step 2: Load and synthesize data
```{r}
#Read and format our lake metadata
lakes <- read.csv("https://pasta-s.lternet.edu/package/data/eml/edi/1029/9/fadd3eaa25b5fdd1fc4efba70e660579") 
lakes_format <- lakes %>%
  dplyr::select(LakeID,Longitude_DD,Latitude_DD) %>%
  rename(Lon=Longitude_DD,Lat=Latitude_DD)%>%
  filter(!is.na(Lon),
         !is.na(Lat))


#Read NC temp
mynetcdf_t <- '../External data/era5.nc' #Downloaded from here: https://cds.climate.copernicus.eu/cdsapp#!/dataset/reanalysis-era5-single-levels-monthly-means?tab=form
#The downloaded data is too large to store on github
ncFile_t <- nc_open(mynetcdf_t)
Lon <- ncvar_get(ncFile_t,"longitude")
Lon <- ifelse(Lon > 180, -(360 - Lon), Lon)
Lat <- ncvar_get(ncFile_t,"latitude")
time <- ncvar_get(ncFile_t,"time")
t_mean <- ncvar_get(ncFile_t,"t2m")
fillvalue <- ncatt_get(ncFile_t, "t2m","missing_value") 
#If you get a vector memory error you may have to follow the steps here
#https://stackoverflow.com/questions/51295402/r-on-macos-error-vector-memory-exhausted-limit-reached
t_mean[t_mean == fillvalue$value] <- NA
months <- month(as.Date("1900-01-01")+hours(time))
years <- year(as.Date("1900-01-01")+hours(time))

#Loop through lakes and add temp data based on lat/lon
combined_t_df <- data.frame(LakeID = rep(lakes_format$LakeID, each = length(months)),
                           Month = rep(months,nrow(lakes_format)),
                           Year = rep(years,nrow(lakes_format)),
                           Temp_C = as.numeric(NA),
                           Lon = as.numeric(NA),
                           Lat = as.numeric(NA)
                           )
for(i in 1:nrow(lakes_format)){
  lon_id <- ifelse(min(abs(lakes_format[i,]$Lon-Lon))<=0.25,which.min(abs(lakes_format[i,]$Lon-Lon)),NA)
  lat_id <- ifelse(min(abs(lakes_format[i,]$Lat-Lat))<=0.25, which.min(abs(lakes_format[i,]$Lat-Lat)),NA)
  if(!is.na(lon_id)&!is.na(lat_id)){
    t_vector1 <- t_mean[lon_id,lat_id,1,] 
    t_vector2 <- t_mean[lon_id,lat_id,2,] 
    t_vector <- t_vector1
    t_vector[is.na(t_vector)] <- t_vector2[is.na(t_vector)]
    combined_t_df$Temp_C[combined_t_df$LakeID==lakes_format[i,]$LakeID]<- t_vector-273.15
    combined_t_df$Lon[combined_t_df$LakeID==lakes_format[i,]$LakeID]<- Lon[lon_id]
    combined_t_df$Lat[combined_t_df$LakeID==lakes_format[i,]$LakeID]<- Lat[lat_id]
  }
}
combined_t_df <- combined_t_df%>%
  mutate(Temp_C = trunc(Temp_C*100000)/100000)
write.csv(combined_t_df,"../Compiled data/historical_temp_output_era5.csv", row.names = F)


#  Precipitation
#Read NC precip
mynetcdf_p <- '../External data/era5_precip.nc' #Downloaded from here: https://cds.climate.copernicus.eu/cdsapp#!/dataset/reanalysis-era5-single-levels-monthly-means?tab=form
ncFile_p <- nc_open(mynetcdf_p)
Lon <- ncvar_get(ncFile_p,"longitude")
Lon <- ifelse(Lon > 180, -(360 - Lon), Lon)
Lat <- ncvar_get(ncFile_p,"latitude")
time <- ncvar_get(ncFile_p,"time")
p_mean <- ncvar_get(ncFile_p,"tp")
fillvalue <- ncatt_get(ncFile_p, "tp","missing_value")
p_mean[p_mean == fillvalue$value] <- NA
months <- month(as.Date("1900-01-01")+hours(time))
years <- year(as.Date("1900-01-01")+hours(time))

#Loop through lakes and add precip data based on lat/lon
combined_p_df <- data.frame(LakeID = rep(lakes_format$LakeID, each = length(months)),
                           Month = rep(months,nrow(lakes_format)),
                           Year = rep(years,nrow(lakes_format)),
                           Total_Precip = as.numeric(NA),
                           Lon = as.numeric(NA),
                           Lat = as.numeric(NA)
                           )
for(i in 1:nrow(lakes_format)){
  lon_id <- ifelse(min(abs(lakes_format[i,]$Lon-Lon))<=0.25,which.min(abs(lakes_format[i,]$Lon-Lon)),NA)
  lat_id <- ifelse(min(abs(lakes_format[i,]$Lat-Lat))<=0.25, which.min(abs(lakes_format[i,]$Lat-Lat)),NA)
  if(!is.na(lon_id)&!is.na(lat_id)){
    p_vector1 <- p_mean[lon_id,lat_id,1,] 
    p_vector2 <- p_mean[lon_id,lat_id,2,] 
    p_vector <- p_vector1
    p_vector[is.na(p_vector)] <- p_vector2[is.na(p_vector)]
    combined_p_df$Total_Precip[combined_p_df$LakeID==lakes_format[i,]$LakeID]<- p_vector
    combined_p_df$Lon[combined_p_df$LakeID==lakes_format[i,]$LakeID]<- Lon[lon_id]
    combined_p_df$Lat[combined_p_df$LakeID==lakes_format[i,]$LakeID]<- Lat[lat_id]
  }
}
write.csv(combined_p_df,"../Compiled data/historical_precip_output_era5.csv", row.names = F)
```
