==> Data analysis/08.5 - Anoxic factor.Rmd <==
---
title: "ASL AF calcs"
author: "Abby Lewis"
date: "2022-08-29"
output: html_document
---

This file calculates anoxic factor for each lake year. 


Step 1: Load data and packages
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(lubridate)
library(rLakeAnalyzer)
library(openair)

#Load data and stratification boundaries. This file is created by "03 - Stratified avgs.Rmd"
full_with_thermo <- read.csv("../Compiled data/Stratified_period_data_with_thermo.csv")%>%
  mutate(Date=as.Date(Date))
#Load lake metadata
lat_long <- read.csv("https://pasta-s.lternet.edu/package/data/eml/edi/1029/9/fadd3eaa25b5fdd1fc4efba70e660579")
```


Step 2: Calculate VW DO
```{r}
#prepare bathymetry
#lake with only one method (data/model)
lake_bats <- read.csv("https://pasta-s.lternet.edu/package/data/eml/edi/1029/9/ba733454c29a5f026efed05a6e6ef75b")%>%
  filter(Area_m2>0)%>%
  group_by(LakeID)%>%
  mutate(methods_n = length(unique(Method)))%>%
  filter(methods_n==1|Method=="model")%>%
  ungroup()

#Check bathymetric data availability
unique(lake_bats$LakeID)[!unique(lake_bats$LakeID)%in%unique(lat_long$LakeID)]
unique(lat_long$LakeID)[!unique(lat_long$LakeID)%in%unique(lake_bats$LakeID)] #missing full bathymetry for 37 lakes

#Total n to start: 609
length(unique(full_with_thermo$LakeID))

#Calculate volume-weighted oxygen concentrations
vw_do <- full_with_thermo%>%
  full_join(lake_bats, by = c("LakeID","Depth_m"))%>%
  filter(!is.na(IntervalVolume_m3))%>%
  filter(Depth_m>hypo_depth)%>%
  mutate(DO_mass = DO_mgL*IntervalVolume_m3,
         Temp_total = Temp_C*IntervalVolume_m3,
         Year = year(Date))%>%
  group_by(Date, LakeID)%>%
  dplyr::summarize(DO_tot = sum(DO_mass),#sum across all hypolimnetic depths
                   vol_tot = sum(IntervalVolume_m3),
                   DO_mgL_vol = DO_tot/vol_tot,#Divide by hypolimnetic volume
                   DO_mgL_area = DO_tot/max(Area_m2),
                   Temp_C = sum(Temp_total)/vol_tot)%>%
  mutate(Year= year(Date))%>%
  group_by(LakeID, Year)%>%
  filter(!is.na(DO_mgL_vol))%>%
  arrange(LakeID, Date)%>%
  mutate(low_point = ifelse(sum(DO_mgL_vol<1)==0,#Identify the lowest DO value for a given year
                            Date[which.min(DO_mgL_vol)],
                            first(Date[DO_mgL_vol<1])))%>%
  filter(is.na(low_point)|Date<=low_point)%>% #Remove days after the lowest DO value
  ungroup()%>%
  mutate(Year = year(Date))

length(unique(vw_do$LakeID)) #581 lakes left
```

Step 3: Model end of stratification date
```{r}
strat_end <- read.csv("../Compiled data/All_data_annual.csv")%>%
  mutate(strat_end_model = 10^(2.62 - 
                           (0.116 * log10(strat_Temp_C_HYPO)) + 
                           (0.042 * log10(MeanDepth_m)) - 
                           (0.002 * Latitude_DD))) 

strat_end %>%
  ggplot(aes(x = strat_end_model))+
  geom_histogram()
```

Step 4: Model start of anoxia date using oxygen demand
```{r}
#Calculate the rate of change in volume-weighted concentrations using lm
DO_THRESH = 1

vw_do_demand <- vw_do%>%
  mutate(Year = year(Date),
         DOY = yday(Date))%>%
  group_by(Year, LakeID)%>%
  dplyr::summarize(n = n(),
            DO_demand_mgLd = -lm(DO_mgL_vol~DOY)$coefficients[2],
            AHOD_mgLd = -lm(DO_mgL_area~DOY)$coefficients[2],
            p_vol = summary(lm(DO_mgL_vol~DOY))$coefficients[8],
            p_area = summary(lm(DO_mgL_area~DOY))$coefficients[8],
            r2_vol = summary(lm(DO_mgL_vol~DOY))$r.squared,
            r2_area = summary(lm(DO_mgL_area~DOY))$r.squared,
            Temp_C = mean(Temp_C, na.rm = T),
            anoxic_onset = (DO_THRESH - lm(DO_mgL_vol~DOY)$coefficients[1]) / 
              lm(DO_mgL_vol~DOY)$coefficients[2])%>%
  filter(n>=3)#Need at least 3 points

#How many rows get removed by filtering so R2>0.5? 246
removed_by_r2 <- vw_do_demand%>%
  filter(r2_vol>.5)

#How many rows get removed by filtering so oxygen demand is positive? 0
removed_by_pos <- removed_by_r2%>%
  filter(DO_demand_mgLd>0)

#Finish QAQC
good_hod <- vw_do_demand%>%
  group_by(LakeID)%>%
  filter(!is.na(DO_demand_mgLd),
         r2_vol>0.5
         )%>%
  dplyr::summarize(n_neg = sum(DO_demand_mgLd<0),
                   n = n(),
                   pct = n_neg/n*100)%>%
  filter(pct<=10)

vw_do_demand_qaqc <- removed_by_pos%>%
  filter(LakeID %in% good_hod$LakeID)

vw_do_demand_qaqc %>%
  ggplot(aes(x = anoxic_onset))+
  geom_histogram()
```

Step 5: Empirically calculate start and end of anoxia date if it can be determined within a certain range (PROXIMITY)
```{r}
PROXIMITY = 7

empirical_anoxic_onset <- vw_do %>%
  group_by(LakeID, Year)%>%
  filter(sum(DO_mgL_vol < DO_THRESH, na.rm = T) >= 1,
         sum(DO_mgL_vol > DO_THRESH, na.rm = T) >= 1) %>%
  mutate(onset = first(Date[DO_mgL_vol < DO_THRESH]),
         before = last(Date[Date<onset]),
         interval = as.numeric(onset-before),
         range = DO_mgL_vol[Date==before] - DO_mgL_vol[Date==onset],
         fraction = (DO_mgL_vol[Date==before] - DO_THRESH) / range)%>%
  filter(interval <= PROXIMITY) %>%
  summarize(anoxic_onset_data = unique(yday(before) + fraction * interval))

empirical_anoxic_end <- full_with_thermo%>%
  full_join(lake_bats, by = c("LakeID","Depth_m"))%>%
  filter(!is.na(IntervalVolume_m3))%>%
  filter(Depth_m > hypo_depth)%>%
  mutate(DO_mass = DO_mgL*IntervalVolume_m3,
         Temp_total = Temp_C*IntervalVolume_m3,
         Year = year(Date))%>%
  group_by(Date, LakeID)%>%
  dplyr::summarize(DO_tot = sum(DO_mass),#sum across all hypolimnetic depths
                   vol_tot = sum(IntervalVolume_m3),
                   DO_mgL_vol = DO_tot/vol_tot) %>%#Divide by hypolimnetic volume
  mutate(Year= year(Date))%>%
  group_by(LakeID, Year)%>%
  filter(!is.na(DO_mgL_vol),
         sum(DO_mgL_vol < DO_THRESH, na.rm = T) >= 1,
         sum(DO_mgL_vol > DO_THRESH, na.rm = T) >= 1) %>%
  mutate(last = last(Date[DO_mgL_vol < DO_THRESH]),
         after = first(Date[Date > last]),
         interval = as.numeric(after - last),
         range = DO_mgL_vol[Date == after] - DO_mgL_vol[Date ==  last],
         fraction = (DO_mgL_vol[Date == after] - DO_THRESH) / range)%>%
  filter(interval <= PROXIMITY) %>%
  summarize(anoxic_end_data = unique(yday(after) - (fraction * interval)))

empirical_strat_end <- full_with_thermo%>%
  full_join(lake_bats, by = c("LakeID","Depth_m"))%>%
  filter(!is.na(IntervalVolume_m3))%>%
  filter(Depth_m > hypo_depth)%>%
  mutate(DO_mass = DO_mgL*IntervalVolume_m3,
         Temp_total = Temp_C*IntervalVolume_m3,
         Year = year(Date))%>%
  group_by(Date, LakeID)%>%
  dplyr::summarize(DO_tot = sum(DO_mass),#sum across all hypolimnetic depths
                   vol_tot = sum(IntervalVolume_m3),
                   DO_mgL_vol = DO_tot/vol_tot, #Divide by hypolimnetic volume
                   mixing_date = unique(Date_end)) %>%
  mutate(Year= year(Date))%>%
  group_by(LakeID, Year)%>%
  filter(Date == last(Date[Date <= as.Date(mixing_date)])) %>%
  mutate(strat_interval = as.numeric(as.Date(mixing_date) - Date),
         mixing_date = yday(mixing_date)) %>%
  filter(strat_interval <= PROXIMITY) %>%
  dplyr::select(LakeID, Year, mixing_date)
```

Plot results
```{r}
anoxic_on_off = vw_do_demand_qaqc %>%
  dplyr::select(Year, LakeID, anoxic_onset) %>%
  rename(anoxic_onset_vhod = anoxic_onset) %>%
  full_join(empirical_anoxic_onset) %>%
  full_join(empirical_anoxic_end) %>%
  full_join(empirical_strat_end) %>%
  full_join(strat_end %>% dplyr::select(LakeID, Year, strat_end_model))

anoxic_on_off %>%
  filter(!is.na(anoxic_onset_data)) %>%
  ggplot(aes(x = anoxic_onset_data, y = anoxic_onset_vhod))+
  geom_abline(slope = 1)+
  geom_smooth(method = "lm")+
  geom_point()

anoxic_on_off %>%
  filter(!is.na(anoxic_end_data)) %>%
  ggplot(aes(x = anoxic_end_data, y = mixing_date))+
  geom_abline(slope = 1)+
  geom_smooth(method = "lm")+
  geom_point()

anoxic_on_off %>%
  filter(!is.na(anoxic_end_data)) %>%
  ggplot(aes(x = anoxic_end_data, y = strat_end_model))+
  geom_abline(slope = 1)+
  geom_smooth(method = "lm")+
  geom_point()

anoxic_on_off %>%
  filter(!is.na(anoxic_end_data)) %>%
  ggplot(aes(x = mixing_date, y = strat_end_model))+
  geom_abline(slope = 1)+
  geom_smooth(method = "lm")+
  geom_point()
```


Step 6: Calculate AF, combining all info
```{r}
lake_bats <- read.csv("https://pasta-s.lternet.edu/package/data/eml/edi/1029/9/ba733454c29a5f026efed05a6e6ef75b")%>%
  filter(Area_m2>0)%>%
  group_by(LakeID)%>%
  mutate(methods_n = length(unique(Method)))%>%
  filter(methods_n==1|Method=="model")%>%
  ungroup()

areas = full_with_thermo %>%
  group_by(LakeID, Year) %>%
  summarize(hypo_depth = round(mean(hypo_depth)),
            SurfaceArea_ha = unique(SurfaceArea_ha)) %>%
  left_join(lake_bats%>% 
              dplyr::select(LakeID, Depth_m, Area_m2), 
            by = c("LakeID",hypo_depth = "Depth_m"))

AF = anoxic_on_off %>%
  mutate(anoxic_onset = ifelse(!is.na(anoxic_onset_data),
                               anoxic_onset_data,
                               anoxic_onset_vhod),
         anoxic_end = strat_end_model,
         duration = anoxic_end - anoxic_onset) %>%
  left_join(areas) %>%
  mutate(AF = duration*Area_m2/(SurfaceArea_ha*10000),
         AF = ifelse(AF < 0, 0, AF))
```


Step 6: If VW DO is above threshold all summer and sampling is done in late summer, AF = 0 (filling in lakes where we couldn't use oxygen demand)
```{r}
summer <- read_csv("../Compiled data/summer_averages_wi.csv")
summer_af <- summer %>%
  group_by(LakeID, Year) %>%
  filter(Layer == "HYPO",
         DO_mgL > 5) %>%
  mutate(AF_summer = 0) %>%
  dplyr::select(LakeID, Year, AF_summer)
  
AF_output <- AF %>%
  dplyr::select(LakeID, Year, AF) %>%
  full_join(summer_af) %>%
  mutate(AF = ifelse(is.na(AF), AF_summer, AF)) %>%
  dplyr::select(LakeID, Year, AF) %>%
  filter(!is.na(AF))

unique(AF_output$LakeID)

write.csv(AF, "../Compiled data/AF.csv", row.names = F)
```
